Divide report for hourly agent. #37
